git-status: do not suggest "e!"
`git status` prints file names relative to the current directory.
Thus, it is possible for a separate script to do this job.

One benefit of relying on `git status` is that it performs a type of
escaping. However, `ex` cannot take advantage of this feature because
it does not appear to support using the "edit" command with a filename
containing a space.

The other reason is that a string can be escaped differently. \n,
$'\n', double quotes, raw characters.